feat: 스켈레톤 지연 렌더링 커스텀 훅 구현 [TM-132]
