+++ src/entities/notice/ui/NoticeList.tsx
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 import useNoticeQueries from '@/entities/notice/model/useNoticeQueries';
 import Skeleton from '@/shared/components/skeleton/Skeleton';
 import { NoticeItem } from './NoticeItem';
+import useDelayPendingState from '@/shared/hooks/useDelayPendingState';
 
 function NoticeList({ isOpen }: { isOpen: boolean }) {
   const { useNoticeListQuery } = useNoticeQueries();
@@ -12,9 +13,11 @@ function NoticeList({ isOpen }: { isOpen: boolean }) {
     isSuccess,
   } = useNoticeListQuery(isOpen); // 공지 리스트 불러오기
 
+  const showSkeleton = useDelayPendingState(1000, isPending); // 1초 후에 스켈레톤 렌더링
+
   return (
     <Container>
-      {isPending &&
+      {showSkeleton &&
         Array.from({ length: 4 }).map((_, idx) => (
           <Skeleton key={`notice-skeleton-${idx}`} width={498} height={58} />
         ))}
